iafactory-academy/backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation & Settings
email-validator>=2.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Payments
stripe==8.0.0

# Email
sendgrid==6.11.0

# AWS S3
boto3==1.34.26
botocore==1.34.26

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-slugify==8.0.1
Pillow==10.2.0
PyPDF2==3.0.1

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# Excel/CSV
openpyxl==3.1.2
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.1.0

# Linting & Formatting
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0

# RAG & LLM (commented out for faster initial build)
# tiktoken>=0.7.0
# unstructured>=0.15.0
markdown>=3.6
beautifulsoup4>=4.12.0
lxml>=5.0.0




pydantic==2.8.2
langchain-core==0.3.30
langchain-community==0.3.13
langchain-openai==0.2.14
langchain==0.3.13
